Make CallSpec an interface.
This divides the public api of CallSpec and it's implementation into two
different classes. This will provided better testability support as well
as better overview of the libraries public api.
